virtio/net: move work to a separate thread
Move the work to a separate thread. This reduces the operation latency,
minimizes the noise between devices and allows us to implement more
aggressive optimization strategies.

While there, also adopt the EVENT_IDX to reduce the notification traffic
between the guest and the host.
