+++ src/devices/src/virtio/net/mod.rs
@@ -13,9 +13,9 @@ pub const TX_INDEX: usize = 1;
 
 mod backend;
 pub mod device;
-pub mod event_handler;
 mod gvproxy;
 mod passt;
+mod worker;
 
 pub use self::device::Net;
 #[derive(Debug)]
