+++ src/vmm/src/builder.rs
@@ -611,12 +611,7 @@ pub fn build_microvm(
     }
 
     #[cfg(feature = "net")]
-    attach_net_devices(
-        &mut vmm,
-        vm_resources.net_builder.iter(),
-        event_manager,
-        intc,
-    )?;
+    attach_net_devices(&mut vmm, vm_resources.net_builder.iter(), intc)?;
 
     if let Some(s) = &vm_resources.boot_config.kernel_cmdline_epilog {
         vmm.kernel_cmdline.insert_str(s).unwrap();
@@ -1236,7 +1231,6 @@ fn attach_console_devices(
 fn attach_net_devices<'a>(
     vmm: &mut Vmm,
     net_devices: impl Iterator<Item = &'a Arc<Mutex<Net>>>,
-    event_manager: &mut EventManager,
     intc: Option<Arc<Mutex<Gic>>>,
 ) -> Result<(), StartMicrovmError> {
     for net_device in net_devices {
@@ -1245,10 +1239,6 @@ fn attach_net_devices<'a>(
         if let Some(ref intc) = intc {
             net_device.lock().unwrap().set_intc(intc.clone());
         }
-
-        event_manager
-            .add_subscriber(net_device.clone())
-            .map_err(StartMicrovmError::RegisterEvent)?;
 
         attach_mmio_device(
             vmm,
